% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDifferentialGeneCoexpression.r
\name{geoDataDownload}
\alias{geoDataDownload}
\title{Function that downloads gene expression data from GEO, after checking the connection}
\usage{
geoDataDownload(GSE_code, verbose = FALSE)
}
\arguments{
\item{GSE_code}{GEO code dataset}

\item{verbose}{prints all the intermediate message to standard output or not}
}
\value{
a gene set gene expression AnnotationDataFrame
}
\description{
Function that downloads gene expression data from GEO, after checking the connection
}
