% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDifferentialGeneCoexpression.r
\name{geoPlatformAnnotationsDownload}
\alias{geoPlatformAnnotationsDownload}
\title{Function that downloads the annotations of a GEO platform}
\usage{
geoPlatformAnnotationsDownload(platformID, verbose = FALSE)
}
\arguments{
\item{platformID}{GEO platform ID}

\item{verbose}{prints all the intermediate message to standard output or not}
}
\value{
a dataframe containing the annotations of the GEO platform
}
\description{
Function that downloads the annotations of a GEO platform
}
