% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDifferentialGeneCoexpression.r
\name{probesetRetrieval}
\alias{probesetRetrieval}
\title{Function that reads a CSV file of probesets or gene symbols and, in the latter case, it retrieves the original probesets}
\usage{
probesetRetrieval(
  probesets_or_gene_symbols,
  csv_file_name,
  platformCode,
  verbose = FALSE
)
}
\arguments{
\item{probesets_or_gene_symbols}{flag saying if we're reading probesets or gene symbols}

\item{csv_file_name}{complete name of CSV file containing the probesets or the gene symbols}

\item{platformCode}{code of the microarray platform for which the probeset-gene symbol mapping should be done}

\item{verbose}{prints all the intermediate message to standard output or not}
}
\value{
a vector of probesets
}
\description{
Function that reads a CSV file of probesets or gene symbols and, in the latter case, it retrieves the original probesets
}
